cmd/compile/internal: Fix float32 constant generation
Fixes #13

math.Float32bits returns a uint32, and then a cast to int64 used to convert
it into a 64bit immediate which overflowed the signed int32 bounds.
Casting it into a signed int32 first forces the compiler to sign extend
the value returned by math.Float32bits which correctly generates the 32bit
signed immediate for the MOV instruction.

+++ src/cmd/compile/internal/riscv/ssa.go
@@ -288,7 +288,7 @@ func ssaGenValue(s *gc.SSAGenState, v *ssa.Value) {
 		// Convert the float to the equivalent integer literal so we can
 		// move it using existing infrastructure.
 		p.From.Type = obj.TYPE_CONST
-		p.From.Offset = int64(math.Float32bits(float32(math.Float64frombits(uint64(v.AuxInt)))))
+		p.From.Offset = int64(int32(math.Float32bits(float32(math.Float64frombits(uint64(v.AuxInt))))))
 		p.To.Type = obj.TYPE_REG
 		p.To.Reg = v.Reg()
 	case ssa.OpRISCVMOVaddr:
